requirements.txt
discord_py_interactions==5.10.0
python-decouple==3.8
pymongo==4.6.0